requirements.txt
django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==22.0.0
whitenoise==6.6.0